refactor: or-1277 split mapping logic into seperate class
