+++ src/AssociationRegistry.Public.Api/Constants/WellknownFacets.cs
@@ -1,0 +1,6 @@
+﻿namespace AssociationRegistry.Public.Api.Constants;
+
+public static class WellknownFacets
+{
+    public const string HoofdactiviteitenCountAggregateName = "Count_By_Hoofdactiviteit";
+}
