+++ src/AssociationRegistry.Public.Api/SearchVerenigingen/Examples/SearchVerenigingenResponseExamples.cs
@@ -7,6 +7,13 @@ using Swashbuckle.AspNetCore.Filters;
 
 public class SearchVerenigingenResponseExamples : IExamplesProvider<SearchVerenigingenResponse>
 {
+    private readonly AppSettings _appSettings;
+
+    public SearchVerenigingenResponseExamples(AppSettings appSettings)
+    {
+        _appSettings = appSettings;
+    }
+
     public SearchVerenigingenResponse GetExamples()
         => new()
         {
@@ -31,7 +38,7 @@ public class SearchVerenigingenResponseExamples : IExamplesProvider<SearchVereni
                     ),
                     ImmutableArray.Create(
                         new Activiteit(123, "Badminton")),
-                    new VerenigingLinks(new Uri("https://???/verenigingen/v123456"))),
+                    new VerenigingLinks(new Uri($"{_appSettings.AssociationRegistryUri}verenigingen/v123456"))),
                 new Vereniging(
                     "V765432",
                     "FWA De Bron",
@@ -51,7 +58,7 @@ public class SearchVerenigingenResponseExamples : IExamplesProvider<SearchVereni
                             "Gent")),
                     ImmutableArray.Create(
                         new Activiteit(456, "Tennis")),
-                    new VerenigingLinks(new Uri("https://???/verenigingen/v765432"))
+                    new VerenigingLinks(new Uri($"{_appSettings.AssociationRegistryUri}verenigingen/v765432"))
                 )),
             Facets = new Facets
             {
